backend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
coreschema==0.0.4
Django==5.0.6
django-cors-headers==4.3.1
django-debug-toolbar==4.4.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
drf-spectacular==0.27.2
ecdsa==0.19.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
pillow==10.3.0
pyasn1==0.6.0
PyJWT==2.8.0
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
rsa==4.9
setuptools==70.1.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
